// src/sites/Site.d.ts
export interface Site {
  URL: string
  REFERENCE_SELECTOR: string | null
  CASE_OFFICER_SELECTOR: string | null
  STATUS_SELECTOR: string | null
  PROPOSAL_SELECTOR: string | null
  RECEIVED_DATE_SELECTOR: string | null
  DECISION_SELECTOR: string | null
  DATE_VALIDATED_SELECTOR: string | null
  ISSUE_DATE_SELECTOR: string | null
  APPLICATION_TYPE_SELECTOR: string | null
  ADDRESS_SELECTOR: string | null
  UPRN_SELECTOR: string | null
  EASTING_SELECTOR: string | null
  NORTHING_SELECTOR: string | null
  WARD_SELECTOR: string | null
  PARISH_SELECTOR: string | null
}